+++ public/kibana-integrations/loader/embedded_visualize_handler.ts
@@ -17,9 +17,10 @@
  * under the License.
  */
 
+ 
 // @ts-ignore
 import { EventEmitter } from 'events';
-import { debounce, forEach, get } from 'lodash';
+import { debounce, forEach, get, isEqual } from 'lodash';
 import * as Rx from 'rxjs';
 import { share } from 'rxjs/operators';
 // @ts-ignore
@@ -48,20 +49,19 @@ import { RequestHandlerParams, Vis } from 'ui/vis';
 import { VisFiltersProvider } from 'ui/vis/vis_filters';
 // @ts-ignore
 import { PipelineDataLoader } from 'ui/visualize/loader/pipeline_data_loader';
-import { visualizationLoader } from './visualization_loader';
+import { visualizationLoader } from 'ui/visualize/loader';
 import { VisualizeDataLoader } from './visualize_data_loader';
+import { onlyDisabledFiltersChanged } from 'plugins/data';
 
-// @ts-ignore
 import { DataAdapter, RequestAdapter } from 'ui/inspector/adapters';
 
-// @ts-ignore
 import { getTableAggs } from 'ui/visualize/loader/pipeline_helpers/utilities';
 import {
   VisResponseData,
   VisSavedObject,
   VisualizeLoaderParams,
   VisualizeUpdateParams,
-} from './types';
+} from 'ui/visualize/loader';
 // @ts-ignore
 import { queryGeohashBounds } from 'ui/visualize/loader/utils';
 
@@ -119,14 +119,13 @@ export class EmbeddedVisualizeHandler {
   private actions: any = {};
   private events$: Rx.Observable<any>;
   private autoFetch: boolean;
-  private errorHandler: any;
+  private abortController?: AbortController;
+  private autoRefreshFetchSubscription: Rx.Subscription | undefined;
 
   constructor(
     private readonly element: HTMLElement,
     savedObject: VisSavedObject,
-    params: EmbeddedVisualizeHandlerParams,
-    injector,
-    errorHandler
+    params: EmbeddedVisualizeHandlerParams
   ) {
     const { searchSource, vis } = savedObject;
 
@@ -139,10 +138,8 @@ export class EmbeddedVisualizeHandler {
       query,
       autoFetch = true,
       pipelineDataLoader = false,
-      Private
+      Private,
     } = params;
-
-    this.errorHandler = errorHandler;
 
     this.dataLoaderParams = {
       searchSource,
@@ -186,10 +183,10 @@ export class EmbeddedVisualizeHandler {
     this.vis.on('reload', this.reload);
     this.uiState.on('change', this.onUiStateChange);
     if (autoFetch) {
-      timefilter.on('autoRefreshFetch', this.reload);
+      this.autoRefreshFetchSubscription = timefilter.getAutoRefreshFetch$().subscribe(this.reload);
     }
 
-    // This is a workaround to give maps visualizations access to data in the
+    // This is a hack to give maps visualizations access to data in the
     // globalState, since they can no longer access it via searchSource.
     // TODO: Remove this as a part of elastic/kibana#30593
     this.vis.API.getGeohashBounds = () => {
@@ -199,9 +196,8 @@ export class EmbeddedVisualizeHandler {
       });
     };
 
-    this.dataLoader = pipelineDataLoader
-      ? new PipelineDataLoader(vis)
-      : new VisualizeDataLoader(vis, Private);
+    //WAZUH disable pipelineLoader
+    this.dataLoader = new VisualizeDataLoader(vis, Private);
     const visFilters: any = Private(VisFiltersProvider);
     this.renderCompleteHelper = new RenderCompleteHelper(element);
     this.inspectorAdapters = this.getActiveInspectorAdapters();
@@ -280,9 +276,10 @@ export class EmbeddedVisualizeHandler {
    */
   public destroy(): void {
     this.destroyed = true;
+    this.cancel();
     this.debouncedFetchAndRender.cancel();
     if (this.autoFetch) {
-      timefilter.off('autoRefreshFetch', this.reload);
+      if (this.autoRefreshFetchSubscription) this.autoRefreshFetchSubscription.unsubscribe();
     }
     this.vis.removeListener('reload', this.reload);
     this.vis.removeListener('update', this.handleVisUpdate);
@@ -455,7 +452,14 @@ export class EmbeddedVisualizeHandler {
     this.fetchAndRender();
   };
 
+  private cancel = () => {
+    if (this.abortController) this.abortController.abort();
+  };
+
   private fetch = (forceFetch: boolean = false) => {
+    this.cancel();
+    this.abortController = new AbortController();
+    this.dataLoaderParams.abortSignal = this.abortController.signal;
     this.dataLoaderParams.aggs = this.vis.getAggConfig();
     this.dataLoaderParams.forceFetch = forceFetch;
     this.dataLoaderParams.inspectorAdapters = this.inspectorAdapters;
@@ -464,23 +468,27 @@ export class EmbeddedVisualizeHandler {
     this.vis.requestError = undefined;
     this.vis.showRequestError = false;
 
-    return this.dataLoader
-      .fetch(this.dataLoaderParams)
-      .then(data => {
-        // Pipeline responses never throw errors, so we need to check for
-        // `type: 'error'`, and then throw so it can be caught below.
-        // TODO: We should revisit this after we have fully migrated
-        // to the new expression pipeline infrastructure.
-        if (data && data.type === 'error') {
-          throw data.error;
-        }
+    return (
+      this.dataLoader
+        // Don't pass in this.dataLoaderParams directly because it may be modified async in another
+        // call to fetch before the previous one has completed
+        .fetch({ ...this.dataLoaderParams })
+        .then(data => {
+          // Pipeline responses never throw errors, so we need to check for
+          // `type: 'error'`, and then throw so it can be caught below.
+          // TODO: We should revisit this after we have fully migrated
+          // to the new expression pipeline infrastructure.
+          if (data && data.type === 'error') {
+            throw data.error;
+          }
 
-        if (data && data.value) {
-          this.dataSubject.next(data.value);
-        }
-        return data;
-      })
-      .catch(this.handleDataLoaderError);
+          if (data && data.value) {
+            this.dataSubject.next(data.value);
+          }
+          return data;
+        })
+        .catch(this.handleDataLoaderError)
+    );
   };
 
   /**
@@ -490,6 +498,9 @@ export class EmbeddedVisualizeHandler {
    * frequently encountered by users.
    */
   private handleDataLoaderError = (error: any): void => {
+    // If the data loader was aborted then no need to surface this error in the UI
+    if (error && error.name === 'AbortError') return;
+
     // TODO: come up with a general way to cancel execution of pipeline expressions.
     if (this.dataLoaderParams.searchSource && this.dataLoaderParams.searchSource.cancelQueued) {
       this.dataLoaderParams.searchSource.cancelQueued();
@@ -499,11 +510,12 @@ export class EmbeddedVisualizeHandler {
     this.vis.showRequestError =
       error.type && ['NO_OP_SEARCH_STRATEGY', 'UNSUPPORTED_QUERY'].includes(error.type);
 
-
-    //Do not show notification toast if it's already being shown a similar toast
-    this.errorHandler.handle(error.message, i18n.translate('common.ui.visualize.dataLoaderError', {
-      defaultMessage: 'Error in visualization',
-    }));
+    toastNotifications.addDanger({
+      title: i18n.translate('common.ui.visualize.dataLoaderError', {
+        defaultMessage: 'Error in visualization',
+      }),
+      text: error.message,
+    });
   };
 
   private rendererProvider = (response: VisResponseData | null) => {