+++ CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to the Wazuh app project will be documented in this file.
 
 
+## Wazuh v3.10.2 - Kibana v7.4.2 - Revision 549
+
+### Added
+
+- Support for Kibana v7.4.2
+
+## Wazuh v3.10.2 - Kibana v7.4.1 - Revision 548
+
+### Added
+
+- Support for Kibana v7.4.1
+
+## Wazuh v3.10.2 - Kibana v7.4.0 - Revision 547
+
+### Added
+
+- Support for Kibana v7.4.0
+- Support for Wazuh v3.10.2.
+
 
 ## Wazuh v3.10.2 - Kibana v7.3.2 - Revision 546